# src/main/webapp/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Smart Agriculture Nutrition",
    "description": "REST API connecting smart agriculture data with personal nutrition tracking. This API bridges the gap between agricultural production and personal health outcomes, creating a complete food-to-health ecosystem.",
    "contact": {
      "name": "Smart Agriculture Team",
      "url": "https://github.com/agriculture-nutrition/api",
      "email": "agriculture@nutrition.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/SmartAgricultureNutrition/api/v1",
      "description": "Development Server"
    }
  ],
...
  "paths": {
    "/crop-nutrition-profiles": {
      "get": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Get all crop nutrition profiles",
        "description": "Retrieve a list of crop nutrition profiles with optional filtering by crop type, growing method, region, or season. Returns comprehensive agricultural data including sustainability scores and nutritional content.",
        "parameters": [
          {
            "name": "crop_type",
            "in": "query",
            "description": "Filter by crop type (e.g., tomato, carrot, wheat)",
            "schema": { "type": "string" },
            "example": "tomato"
          },
          {
            "name": "growing_method",
            "in": "query",
            "description": "Filter by growing method (e.g., organic, hydroponic, traditional)",
            "schema": { "type": "string" },
            "example": "organic"
          }
        ],
        "responses": {
          "200": {
            "description": "List of crop nutrition profiles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CropNutritionProfile"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Create new crop nutrition profile",
        "description": "Create a new crop nutrition profile with automatic sustainability scoring and nutritional prediction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CropNutritionProfile" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crop profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropNutritionProfile"
                }
              }
            }
          }
        }
      }
    },
    "/crop-nutrition-profiles/{profileId}": {
      "get": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Get specific crop nutrition profile",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Crop profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropNutritionProfile"
                }
              }
            }
          },
          "404": { "description": "Crop profile not found" }
        }
      },
      "put": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Update crop nutrition profile",
        "description": "Update an existing crop nutrition profile with recalculated sustainability scoring",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CropNutritionProfile" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Crop profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropNutritionProfile"
                }
              }
            }
          },
          "404": { "description": "Crop profile not found" }
        }
      },
      "delete": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Delete crop nutrition profile",
        "description": "Remove a crop nutrition profile and all associated harvest batches",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Crop profile deleted successfully" },
          "404": { "description": "Crop profile not found" }
        }
      }
    },
    "/crop-nutrition-profiles/{profileId}/harvest-batches": {
      "get": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Get harvest batches for crop profile",
        "description": "Retrieve harvest batches with quality grading and actual nutrient analysis",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "harvest_date",
            "in": "query",
            "description": "Filter by harvest date (YYYY-MM-DD)",
            "schema": { "type": "string" },
            "example": "2024-03-15"
          },
          {
            "name": "quality_grade",
            "in": "query",
            "description": "Filter by quality grade (A, B, C)",
            "schema": { "type": "string" },
            "example": "A"
          }
        ],
        "responses": {
          "200": {
            "description": "Harvest batches retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/HarvestBatch" }
                }
              }
            }
          },
          "404": { "description": "Crop profile not found" }
        }
      },
      "post": {
        "tags": ["Crop Nutrition Profiles"],
        "summary": "Add harvest batch to crop profile",
        "description": "Add a new harvest batch with automatic nutrient analysis based on growing conditions",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HarvestBatch" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Harvest batch created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HarvestBatch" }
              }
            }
          },
          "404": { "description": "Crop profile not found" }
        }
      }
    },
...
    "/nutrition/crop/{cropName}": {
      "get": {
        "tags": ["Nutrition"],
        "summary": "Get nutrition data for a crop",
        "description": "Get comprehensive nutrition data from USDA FoodData Central",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "tomato"
          }
        ],
        "responses": {
          "200": {
            "description": "Nutrition data retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FoodNutritionData" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CropNutritionProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "cropName": {
            "type": "string",
            "example": "Organic Cherry Tomatoes"
          },
          "farmLocation": { "type": "string", "example": "Jyväskylä, Finland" },
          "growingMethod": { "type": "string", "example": "Organic" },
          "sustainabilityScore": {
            "type": "number",
            "format": "double",
            "example": 8.7
          },
          "soilNutrients": { "type": "object" },
          "cropNutrients": { "type": "object" },
          "certifications": { "type": "array", "items": { "type": "string" } },
          "links": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Link" }
          }
        }
      },
...
      "FoodNutritionData": {
        "type": "object",
        "properties": {
          "fdcId": { "type": "integer", "format": "int64" },
          "description": { "type": "string" },
          "foodNutrients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "amount": { "type": "number", "format": "double" },
                "unitName": { "type": "string" },
                "nutrient": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "unitName": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
      "HarvestBatch": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "cropProfileId": { "type": "integer", "format": "int64" },
          "harvestDate": { "type": "string", "example": "2024-03-15" },
          "quantityKg": {
            "type": "number",
            "format": "double",
            "example": 150.5
          },
          "qualityGrade": { "type": "string", "example": "A" },
          "actualNutrients": { "type": "object" },
          "storageConditions": {
            "type": "string",
            "example": "Cold storage at 4°C"
          },
          "links": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Link" }
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "link": { "type": "string" },
          "rel": { "type": "string" }
        }
      }
    }
  }
}